requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
numpy==1.26.4
opencv-python==4.10.0.84

